telecom-pipeline/data-generator/requirements.txt
kafka-python==2.0.2
faker==18.11.2
